twinkletaps-be/pyproject.toml
[tool.poetry]
name = "twinkletaps-be"
version = "0.1.0"
description = "Twinkletaps Backend"
authors = ["Mateusz Krainski <mateusz@krainski.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
requests = "^2.32.3"
pytest = "^8.2.2"
pytest-django = "^4.8.0"


[tool.poetry.group.dev.dependencies]
darglint = "^1.8.1"
flake8 = "^7.1.0"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^24.4.26"
flake8-builtins = "^2.5.0"
flake8-comprehensions = "^3.14.0"
flake8-docstrings = "^1.7.0"
flake8-eradicate = "^1.5.0"
flake8-isort = "^6.1.1"
flake8-spellcheck = "^0.28.0"
pep8-naming = "^0.14.1"
flake8-pyproject = "^1.2.3"
black = "^24.4.2"
flake8-black = "^0.3.6"
isort = "^5.13.2"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
dictionaries = ["en_US", "python", "technical", "django"]
builtins-ignorelist = ["id"]
ignore = ["D104", "D100", "D101", "D106"]
per-file-ignores = """
  **/migrations/*: E501
"""

[tool.isort]
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
profile = "black"

[tool.poe]
envfile = ".env"

[tool.poe.tasks]
configure = "./scripts/configure.sh"
manage = "python ./scripts/manage.py"
lint = "flake8 twinkletaps_be settings scripts"
generate-secret = 'python -c "import secrets; print(secrets.token_urlsafe())"'
test = 'pytest --showlocals'

[tool.poe.tasks.test-full]
sequence = [
  { cmd = "./scripts/try-runserver.sh" },
  { ref = "test" }
]